fix: Refactor to use the gem implementation convention
