+++ .gitignore
@@ -8,3 +8,4 @@
 /tmp/
 *.gem
 .env
+.Idea/
