+++ lib/types_export.rb
@@ -1297,4 +1297,4 @@ require_relative "merge_ruby_client/accounting/types/vendor_credit_request"
 require_relative "merge_ruby_client/accounting/types/vendor_credit_response"
 require_relative "merge_ruby_client/accounting/types/warning_validation_problem"
 require_relative "merge_ruby_client/accounting/types/webhook_receiver"
-require_relative "merge_ruby_client/integrations/types/integration"
+require_relative "merge_ruby_client/integrations/types/provider"
